coalesce logic from output-collection to dist-collection/index.ts
